compiler: Use Array[Byte] to represent bytevectors
We were using Array[Short] to emulate unsigned bytes. This seems a bit
hacky and wastes memory. Instead, manually zero extend in the two places
we need to treat the bytes as unsigned.

+++ compiler/src/main/scala/AbstractSyntaxTree.scala
@@ -176,7 +176,7 @@ case class Vector(elements: ScalaVector[Datum]) extends Datum {
     "#(" + elements.map(_.toString).mkString(" ") + ")"
 }
 
-case class Bytevector(elements: ScalaVector[Short]) extends NonSymbolLeaf {
+case class Bytevector(elements: ScalaVector[Byte]) extends NonSymbolLeaf {
   val schemeType = vt.BytevectorType
 
   override def toString =
