+++ compiler/src/main/scala/planner/intermediatevalue/ConstantValue.scala
@@ -137,7 +137,7 @@ case class ConstantBooleanValue(value: Boolean) extends TrivialConstantValue(ct.
   }
 }
 
-case class ConstantBytevectorValue(elements: Vector[Short]) extends TrivialConstantValue(ct.BytevectorCell, elements, ps.CreateBytevectorCell.apply) with BoxedOnlyValue with KnownBytevector {
+case class ConstantBytevectorValue(elements: Vector[Byte]) extends TrivialConstantValue(ct.BytevectorCell, elements, ps.CreateBytevectorCell.apply) with BoxedOnlyValue with KnownBytevector {
   val typeDescription = "constant bytevector"
   val bytevectorLength = elements.size
 }
